feat: move gen safe script to dapp dir
