+++ apps/dapp/package.json
@@ -8,7 +8,7 @@
     "typechain": "typechain --target=ethers-v5 --out-dir src/types/typechain --glob abi/**/*.json",
     "prepare": "yarn typechain:compile && yarn typechain:safe && yarn typechain:copy && yarn codegen:safe-api-v1 && yarn typechain",
     "typechain:compile": "cd ../../protocol && yarn compile --force",
-    "typechain:safe": "cd ../../protocol && yarn typechain:safe",
+    "typechain:safe": "node ./src/safe/sdk/scripts/generateTypechainFiles.mjs && cp -r typechain/safe src/types/typechain && rm -rf typechain",
     "typechain:copy": "cp -r ../../protocol/typechain ./src/types",
     "dev": "vite",
     "compile": "tsc",
@@ -19,16 +19,14 @@
     "version:release": "standard-version",
     "version:prerelease": "standard-version --prerelease beta",
     "build-with-stats": "tsc && BUILD_STATS=true vite build",
-    "safe-api-v1-host":"(cd src/safe/open-api && ./replace-host.sh && rm -f apiv1.jsone)",
+    "safe-api-v1-host": "(cd src/safe/open-api && ./replace-host.sh && rm -f apiv1.jsone)",
     "codegen:safe-api-v1": "yarn safe-api-v1-host && openapi --input ./src/safe/open-api/apiV1.json --output ./src/safe/open-api/client"
   },
   "dependencies": {
     "@balancer-labs/sdk": "^1.0.4",
-    "@gnosis.pm/safe-contracts": "^1.3.0",
     "@safe-global/safe-apps-provider": "^0.18.0",
     "@safe-global/safe-apps-sdk": "^8.1.0",
     "@safe-global/safe-core-sdk-types": "2.3.0",
-    "@safe-global/safe-deployments": "^1.26.0",
     "@tanstack/react-query": "^4.36.1",
     "@tippyjs/react": "^4.2.6",
     "@web3-onboard/coinbase": "^2.2.5",
@@ -70,6 +68,7 @@
     "@babel/core": "^7.16.0",
     "@commitlint/cli": "^15.0.0",
     "@commitlint/config-conventional": "^15.0.0",
+    "@gnosis.pm/safe-contracts": "^1.3.0",
     "@nomiclabs/hardhat-ethers": "2.2.1",
     "@synthetixio/synpress": "^1.2.0",
     "@typechain/ethers-v5": "10.1.0",
